requirements.txt
flask-socketio
WSGIserver
gevent
gevent-websocket
flask_cors
paho-mqtt
redis
eventlet